requirements.txt
asgiref==3.5.2
awsebcli==3.20.3
botocore==1.23.54
cement==2.8.2
certifi==2022.6.15
charset-normalizer==2.0.12
cloudinary==1.29.0
colorama==0.4.3
Django==4.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
future==0.16.0
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
pathspec==0.9.0
Pillow==9.2.0
pypiwin32==223
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==304
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
tzdata==2022.2
urllib3==1.26.12
wcwidth==0.1.9
whitenoise==6.2.0